Move sequence oidc tests
